refactor: Extracted function (renameComponent)
