+++ src/steps/create-registries/index.ts
@@ -1,5 +1,3 @@
 export * from './create-registry.js';
-export * from './get-base-component-name.js';
 export * from './has-registry.js';
-export * from './pass-component-name-to-base-component.js';
-export * from './update-references.js';
+export * from './rename-component.js';
